[ALPS05144979] ISP: fix code defect
[Detail] Fix code defect to prevent overflow.

MTK-Commit-Id: dc8d706971019835d639d76c5c9a4dc4757b7bcc

CR-Id: ALPS05144979
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/misc/mediatek/cameraisp/src/mt6761/camera_isp.c
@@ -12998,7 +12998,7 @@ static signed int mmap_kmem(struct file *filp, struct vm_area_struct *vma)
 static signed int ISP_mmap(struct file *pFile, struct vm_area_struct *pVma)
 {
 	unsigned long length = 0;
-	unsigned int pfn = 0x0;
+	unsigned long pfn = 0x0;
 
 	log_dbg("- E.");
 	length = (pVma->vm_end - pVma->vm_start);
